+++ js/const.js
@@ -1,0 +1,37 @@
+export const DESCRIPTIONS = [
+  'Поездка по Санкт-Питербургу',
+  'Золотая осень',
+  'Моя любимая собака',
+  'Рыбалка это моё хобби',
+  'Поездка на Волгу',
+  'Поездка на море',
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const NAMES = [
+  'Владимир',
+  'Елена',
+  'Ксения',
+  'Владислав',
+  'Егор',
+  'Сергей',
+  'Наталья',
+  'Евгений',
+  'Дарья'
+];
+
+export const MIN_AVATAR = 1;
+export const MAX_AVATAR = 6;
+export const MIN_COMMENTS_COUNT = 0;
+export const MAX_COMMENTS_COUNT = 5;
+export const MIN_LIKES_COUNT = 15;
+export const MAX_LIKES_COUNT = 200;
+export const PHOTOS_COUNT = 25;
