+++ js/util.js
@@ -1,0 +1,33 @@
+//https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Math/random
+
+const getRandomInt = (min, max) => {
+  if (min < 0) {
+    min = 0;
+  }
+  if (max < 0) {
+    max = 0;
+  }
+  if (min > max) {
+    const swap = min;
+    min = max;
+    max = swap;
+  }
+  min = Math.ceil(min);
+  max = Math.floor(max);
+
+  return Math.floor(Math.random() * (max - min)) + min;
+};
+
+const getLengthOfString = (string, maxLength) => string.length <= maxLength;
+
+const getRandomArrayElement = (elements) => elements[getRandomInt(0, elements.length - 1)];
+
+const getRandomArray = (callback, count) => {
+  const data = [];
+  for (let i = 0; i < count; i++){
+    data.push(callback(i + 1));
+  }
+  return data;
+};
+
+export{getRandomInt, getLengthOfString, getRandomArrayElement, getRandomArray};
